Move ASCforIoT Security message e2e test to a different class
